core/multitenant_v3 default_url_prefix with tenant_id
This changes a few of the integration test to include
the tenant_id functionality.

There are many more integration tests for the mappers that have
not includes tests for the tenant_id functionality.
They however pass use to using `.endswith(` meaning that
the resultant string still ends with the same value.

See: tests/integration/common/artifacts/test_registry_base.py
for an example of this.