Implement gspread `open` identical patch
Reasoning for patching: `gspread`'s `open` sheet function currently
doesn't allow for passing a `folder_id` and only looking for sheets with
the name given within that folder (I don't know why they haven't done
this already). Implementing this is trivial, and would massively reduce
the chance of having sheet name conflicts which would create bugs and
could be hard to diagnose.

This this commit adds a patch with code identical to `gspread`s current
implementation, i.e. a "stub" patch. This enables verifying the patch is
applying correctly and `gspread`s current behaviour is preserved, before
extending the functionality.