FIX: add missing scrape_since_days param to CLI
Also adds a cast to int in the nb.

The naming and code conventions for all this is a bit whacky tbh: we've
got `scrape_since_days' cli param, which is assigned to `SINCE_DAYS` for
facebook posts but needed to be assigned to `SCRAPE_SINCE_DAYS' for
youtube videos.