Add instance type check for tfa `features` output
This fails, i.e. the function isn't doing what is indicated by the
comments within the function as intended - it is returning a DataFrame
instead of a Series. I think this needs looking into/fixing?

This commit also cleans up the test by using `paramterize`. It also uses
`.loc` for value selection.