Implement adding "retweeted_user_name" to tweets
This enabled creating the retweet graph directly from the final data
set. All graphs should go this route if possible; it enhanced the final
data sets with more data, and simplifies the pipelines and removes tons
of code (i.e. we can delete all the legacy retweet data pull and ipynb
for it etc.).