Add print url for Graphistry integration test
Includes note in doc string to use `pytest -s`. This is important as if
you're playing around with different URL param settings, they only show
in the URL it outputs, not if you open the graph from the Graphistry
hub!!! That's pretty bad, so even more of a reason to use the redirect
HTML files only.