Implement handle graph w. disconnected component
This is just a work around for now, as really those community algos
should identify the largest connected sub-component and compute
communities for that sub-graph. PageRank and the others already do that
so, it is annoying some of them don't. We should implement it manually
for those (Spin Glass and Louvian) otherwise we won't ever get those in
the end graphs (because chances are most real world networks will have
disconnected sub-components.